+++ jjava-jupyter/src/test/java/org/dflib/jjava/jupyter/kernel/magic/MagicsRegistryTest.java
@@ -6,7 +6,6 @@ import org.junit.jupiter.api.Test;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
-import java.util.Map;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -14,7 +13,7 @@ public class MagicsRegistryTest {
 
     @Test
     public void lineMagic() throws Exception {
-        MagicsRegistry registry = new MagicsRegistry(Map.of("test", (k, args) -> args), Map.of());
+        MagicsRegistry registry = new MagicsRegistry().registerLineMagic("test", (k, args) -> args);
 
         List<String> args = Arrays.asList("arg1", "arg2");
         List<String> out = registry.evalLineMagic(null, "test", args);
@@ -24,11 +23,11 @@ public class MagicsRegistryTest {
 
     @Test
     public void cellMagic() throws Exception {
-        MagicsRegistry registry = new MagicsRegistry(Map.of(), Map.of("test", (k, args, body) -> {
+        MagicsRegistry registry = new MagicsRegistry().registerCellMagic("test", (k, args, body) -> {
             List<String> out = new ArrayList<>(args);
             out.add(body);
             return out;
-        }));
+        });
 
         List<String> args = Arrays.asList("arg1", "arg2");
         String body = "body";
@@ -44,7 +43,6 @@ public class MagicsRegistryTest {
     public void lineCellMagic() throws Exception {
         class Magic implements LineMagic<List<String>, BaseKernel>, CellMagic<List<String>, BaseKernel> {
 
-
             @Override
             public List<String> eval(BaseKernel kernel, List<String> args, String body) {
                 List<String> out = new ArrayList<>(args);
@@ -59,7 +57,9 @@ public class MagicsRegistryTest {
         }
 
         Magic magic = new Magic();
-        MagicsRegistry registry = new MagicsRegistry(Map.of("test", magic), Map.of("test", magic));
+        MagicsRegistry registry = new MagicsRegistry()
+                .registerLineMagic("test", magic)
+                .registerCellMagic("test", magic);
 
         List<String> args = Arrays.asList("arg1", "arg2");
         String body = "body";