+++ jjava-kernel/src/test/java/org/dflib/jjava/kernel/ExtensionLoadingExtraClasspathTest.java
@@ -1,11 +1,9 @@
 package org.dflib.jjava.kernel;
 
+import org.dflib.jjava.jupyter.kernel.util.PathsHandler;
 import org.junit.jupiter.api.Test;
 
 import java.nio.file.Path;
-import java.util.List;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
@@ -19,10 +17,7 @@ class ExtensionLoadingExtraClasspathTest {
                 "java/META-INF/services/org.dflib.jjava.jupyter.Extension"
         );
 
-        List<String> extraClasspath = Stream.of(jar)
-                .map(Path::toAbsolutePath)
-                .map(Path::toString)
-                .collect(Collectors.toList());
+        String extraClasspath = PathsHandler.joinPaths(PathsHandler.resolveGlobs(jar.toAbsolutePath().toString()));
 
         String extensionClassName = "org.dflib.jjava.kernel.test.TestExtension";
         String installationsProperty = extensionInstallationsProperty(extensionClassName);