+++ RELEASE-NOTES.md
@@ -12,7 +12,9 @@
 - #82 Refactoring: Move JavaKernel init logic from constructor to Builder
 - #83 Simplify Kernel metadata loading approach
 - #84 Refactoring: reorg Maven modules for reusability
-
+- #86 ineffecient / incorrect classpath handling
+- #89 Deprecate redundant magics: %jars, %addMavenDependency
+- #90 Deprecate Ivy artifact syntax for %maven magic
 
 ## 1.0-a5